CU-24wgdkg seeder内を記述してデータを生成した
