+++ backend/database/factories/FavoriteCityFactory.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Database\Factories\Hitchhikes\Spots;
+namespace Database\Factories;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-class SuccessFactory extends Factory
+class FavoriteCityFactory extends Factory
 {
     /**
      * Define the model's default state.
@@ -14,8 +14,7 @@ class SuccessFactory extends Factory
     public function definition()
     {
         return [
-            'created_at' => $this->faker->dateTimeThisDecade(),
-            'updated_at' => $this->faker->dateTimeThisDecade(),
+            
         ];
     }
 }
