--- backend/database/factories/Hitchhikes/Spots/HitchhikeResponseFactory.php
@@ -1,22 +1,0 @@
-<?php
-
-namespace Database\Factories\Hitchhikes\Spots;
-
-use Illuminate\Database\Eloquent\Factories\Factory;
-
-class HitchhikeResponseFactory extends Factory
-{
-    /**
-     * Define the model's default state.
-     *
-     * @return array
-     */
-    public function definition()
-    {
-        return [
-            'message' => $this->faker->realText(),
-            'created_at' => $this->faker->dateTimeThisDecade(),
-            'updated_at' => $this->faker->dateTimeThisDecade(),
-        ];
-    }
-}
