+++ backend/database/factories/MemoFactory.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Database\Factories;
+
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+class MemoFactory extends Factory
+{
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'memo' => $this->faker->realText(),
+            'img' => $this->faker->imageUrl(),
+        ];
+    }
+}
